services/ingester/pyproject.toml
[project]
name = "ingester"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "loguru>=0.7.3",
    "paho-mqtt>=2.1.0",
]
